Adjusted file structure for more efficient collaboration. Should be in line with architecture requirements.
