--- src/Presenter/AppPresenter.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import AppView from "../View/AppView";
-
-function App(props){
-    return <AppView />;
-}
-
-export default App;
