--- src/View/ExampleView.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function ExampleView(props){
-    return (<div>Hello World!!</div>);
-};
-
-export default ExampleView;
