+++ src/pages/AppView.js
@@ -1,0 +1,14 @@
+import RegistrationPageView from "./RegistrationPage/RegistrationPageView";
+
+
+function AppView() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <RegistrationPageView/>
+      </header>
+    </div>
+  );
+}
+
+export default AppView;
