+++ src/store/store.js
@@ -1,0 +1,8 @@
+import { configureStore } from "@reduxjs/toolkit";
+import authReducer from "../feature/authReducer";
+
+export const store = configureStore({
+    reducer: {
+        auth: authReducer,
+    },
+});
